# THRIFT-5133: Use ArrayPool when reading and writing strings to improve performance.