Uses ArrayPool to save on buffer allocations when reading/writing strings.

This change is going into 0.13.0, the reason for that is I really need this improvement in https://github.com/open-telemetry/opentelemetry-dotnet which is on 0.13.0 and the master branch has breaking changes which I won't be able to consume easily.

I will also PR into master, depending on feedback from the team.