@Jens-G 2020-03-06
> This change is going into 0.13.0, 
Will it?