@CodeBlanch 2020-03-09
@Jens-G Thanks for the reply! Trying to read between the lines here. So we're saying performance is not a big enough reason to patch and we want users to take on the next major version? I'm fine with that, but master has these breaking changes which make adoption difficult:

```csharp
public abstract class TTransport : IDisposable
...
        public abstract void UpdateKnownMessageSize(long size);
        public abstract void CheckReadBytesAvailable(long numBytes);
```

OT has its own TTransport (for some reason): https://github.com/open-telemetry/opentelemetry-dotnet/blob/master/src/OpenTelemetry.Exporter.Jaeger/Implementation/InMemoryTransport.cs

Is there guidance somewhere on how to implement the `UpdateKnownMessageSize`/`CheckReadBytesAvailable` API?

My hope here was a small patch to 0.13.X for people on the existing API using it in high-volume scenarios and then also put into master for the next major (whenever people decide to upgrade).

Here are some performance numbers with and without this change (see `Allocated` column):

Before:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
  DefaultJob : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT


```
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev |      Gen 0 |     Gen 1 | Gen 2 | **Allocated** | Completed Work Items | Lock Contentions |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|-----------:|----------:|------:|----------:|---------------------:|-----------------:|
| **JaegerExporter_Batching** |               **1** |         **10000** |    **146.5 ms** |   **2.92 ms** |   **4.63 ms** |          **-** |         **-** |     **-** |   **6.74 MB** |               **1.0000** |                **-** |
| **JaegerExporter_Batching** |              **10** |         **10000** |  **1,442.4 ms** |  **74.31 ms** |  **85.58 ms** |  **8000.0000** |         **-** |     **-** |   **65.3 MB** |               **2.0000** |                **-** |
| **JaegerExporter_Batching** |             **100** |         **10000** | **14,192.9 ms** | **216.25 ms** | **180.58 ms** | **81000.0000** | **1000.0000** |     **-** | **650.91 MB** |               **4.0000** |                **-** |

After:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
  DefaultJob : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT


```
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 |  **Allocated** | Completed Work Items | Lock Contentions |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|------:|------:|------:|-----------:|---------------------:|-----------------:|
| **JaegerExporter_Batching** |               **1** |         **10000** |    **153.3 ms** |   **3.43 ms** |   **6.10 ms** |     **-** |     **-** |     **-** |  **256.76 KB** |               **1.0000** |                **-** |
| **JaegerExporter_Batching** |              **10** |         **10000** |  **1,511.4 ms** |  **29.43 ms** |  **41.26 ms** |     **-** |     **-** |     **-** |  **438.62 KB** |               **2.0000** |                **-** |
| **JaegerExporter_Batching** |             **100** |         **10000** | **15,289.7 ms** | **298.87 ms** | **332.20 ms** |     **-** |     **-** |     **-** | **2270.66 KB** |               **4.0000** |                **-** |

That's **KB** not MB!

> and someone who's willing to prepare such a release

Happy to do it, don't know the process.