@Jens-G 2020-03-11
> UpdateKnownMessageSize/CheckReadBytesAvailable API
see THRIFT-5021 and THRIFT-5027

>That's KB not MB!
I agree on the alloc numbers (great!), but why is it 7% slower?

> release process
https://github.com/apache/thrift/blob/master/doc/ReleaseManagement.md