@CodeBlanch 2020-03-12
@Jens-G That release process looks pretty intense! I'm just going to close this and focus on the `master` version. Any rough ETA on when 0.14.0 will release?

> why is it 7% slower?

To be honest, I don't know. Each time I run the benchmark the speed slightly differs. I would expect a slight hit pooling the buffers, but not a big one. Check out the more focused numbers on #2057 and let me know there if you want me to dig into it, they seem more reasonable.