@Jens-G 2020-03-13
Ok, agree, please proceed.

I guess thats the "curse" of modern computing equipment that it is hard to get reproducible numbers beyond 2 significant digits. I had (and have) the same experience with my own tests and profiling sessions acrosss different languages, test routines and profiling tools.